requirements.txt
aiohttp>=3.9.1
pydantic>=2.0
PyGithub>=2.0
python-dotenv>=1.0.0
Jinja2>=3.1.2
schedule>=1.2.0
telethon>=1.30.0
google-generativeai>=0.4.0
requests>=2.28.0
google-auth>=2.22.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
asyncio>=3.4.3
urllib3>=1.26.0